requirements.txt
# NPC Dialogue System - Python Dependencies

# HTTP client for Ollama API
requests>=2.31.0

# API Server (Issue #1 - Unity Integration)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# RAG/Vector Database (Issue #3 - Lore System)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Optional: Enhanced features
python-dotenv>=1.0.0
# websockets>=12.0  # For WebSocket streaming
# python-multipart>=0.0.6  # For file uploads
